+++ src/main/java/truffle/nodes/exps/TFIntNode.java
@@ -4,9 +4,9 @@ import com.oracle.truffle.api.frame.VirtualFrame;
 import truffle.nodes.TFNode;
 
 public class TFIntNode extends TFNode {
-    private final int value;
+    private final long value;
 
-    public TFIntNode(int value) {
+    public TFIntNode(long value) {
         this.value = value;
     }
 
@@ -16,7 +16,7 @@ public class TFIntNode extends TFNode {
     }
 
     @Override
-    public int executeInt(VirtualFrame frame) {
+    public long executeInt(VirtualFrame frame) {
         return this.value;
     }
