+++ src/main/java/truffle/types/Types.java
@@ -2,7 +2,7 @@ package truffle.types;
 
 import com.oracle.truffle.api.dsl.TypeSystem;
 
-@TypeSystem({int.class, boolean.class, TFFunction.class, String.class})
+@TypeSystem({long.class, boolean.class, TFFunction.class, String.class})
 public abstract class Types {
     //TODO type check;
 }
